python/src/pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.15"
dependencies = [
    "pyaudio>=0.2.14",
    "python-dotenv>=1.2.1",
    "gladiaio-sdk @ file:///Users/egenthon/Documents/Code/Gladia-sdk-samples/sdk/packages/sdk-python",
]
